English Dictionary App & Library Book Management System in python
01: English Dictionary App

English Dictionary App that responds with definitions of English words according to the given input. The aim of this project is to create an English Dictionary app that returns definitions of English words.

02: Library Book Management System

Library Book Management System based on given inputs. The Management System will be implemented with the front-end GUI developed using Tkinter and the backend database implemented in SQLite.